refactor: functions 구조 개편 import 라인을 줄이기 위한 구조 개편 1. 유저와 컴퓨터 input 관리 2. ball, strike 관리 3. 게임 결과 관리
