+++ index.html
@@ -13,10 +13,10 @@
         올바른 예) 139 <br />
         틀린 예) 122
       </p>
-      
+      <form>
         <input type="text" id="user-input" />
         <button id="submit">확인</button>
-      
+      </form>
       <h3>📄 결과</h3>
       <div id="result"></div>
     </div>
