--- src/functions/set-ball-strike.js
@@ -1,16 +1,0 @@
-import { state } from '../state.js';
-
-export default function setBallStrike(computerInput, userInput) {
-  const computerInputString = computerInput.toString();
-  const userInputString = userInput.toString();
-  for (let i = 0; i < computerInputString.length; i += 1) {
-    if (
-      userInputString.includes(computerInputString[i]) &&
-      userInputString.indexOf(computerInputString[i]) === i
-    ) {
-      state.strikeCount += 1;
-    } else if (userInputString.includes(computerInputString[i])) {
-      state.ballCount += 1;
-    }
-  }
-}
